+++ line_vins_estimator/src/estimator.h
@@ -18,6 +18,7 @@
 #include "factor/projection_factor.h"
 #include "factor/projection_td_factor.h"
 #include "factor/marginalization_factor.h"
+#include "factor/line_projection_factor.h"
 
 #include <unordered_map>
 #include <queue>